requirements.txt
pandas
python-dateutil
pytz
ipython
notebook